+++ grunt/config/clean.js
@@ -1,0 +1,25 @@
+/**
+ * Generated header by C0ZEN for 21-days project
+ * Generated file clean on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 01/01/2018
+ * Time: 18:50
+ * Version: 1.0.0
+ */
+module.exports = {
+	tmpScripts: {
+		src: [
+			'.tmp/release/scripts.js',
+			'.tmp/release/scripts.js.map',
+			'.tmp/release/21-days.js',
+			'.tmp/release/21-days.js.map'
+		]
+	},
+	releaseScripts: {
+		src: [
+			'docs/scripts/21-days.min.js',
+			'docs/scripts/21-days.min.js.map'
+		]
+	}
+};